internal: Rename slices to sliceutil
Rename 'slices' to 'sliceutil' to allow use of the std 'slices' package
without import conflicts.
