--- internal/slices/doc.go
@@ -1,2 +1,0 @@
-// Package slices contains generic functions for working with slices.
-package slices
