+++ internal/sliceutil/doc.go
@@ -1,0 +1,3 @@
+// Package sliceutil contains generic functions for working with slices.
+// Use it to augment the std "slices" package.
+package sliceutil
