+++ internal/sliceutil/remove_test.go
@@ -1,4 +1,4 @@
-package slices
+package sliceutil
 
 import (
 	"testing"
